@@ -1,12 +1,11 @@
 ﻿using Plugin.LocalNotification.AndroidOption;
 using Plugin.LocalNotification.iOSOption;
 using Plugin.LocalNotification.Json;
-using System;
 
 namespace Plugin.LocalNotification
 {
     /// <summary>
-    /// 
+    ///
     /// </summary>
     public interface ILocalNotificationBuilder
     {
@@ -18,21 +17,21 @@ namespace Plugin.LocalNotification
         ILocalNotificationBuilder AddCategory(NotificationCategory category);
 
         /// <summary>
-        /// 
+        ///
         /// </summary>
         /// <param name="serializer"></param>
         /// <returns></returns>
         ILocalNotificationBuilder SetSerializer(INotificationSerializer serializer);
 
         /// <summary>
-        /// 
+        ///
         /// </summary>
         /// <param name="android"></param>
         /// <returns></returns>
         ILocalNotificationBuilder AddAndroid(Action<IAndroidLocalNotificationBuilder> android);
 
         /// <summary>
-        /// 
+        ///
         /// </summary>
         /// <param name="iOS"></param>
         /// <returns></returns>
