@@ -1,14 +1,8 @@
-﻿#if NET6_0_OR_GREATER
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
-using Microsoft.Maui.ApplicationModel;
-using Microsoft.Maui.Hosting;
+﻿using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Maui.LifecycleEvents;
-using System;
-using System.Linq;
+
 #if WINDOWS
 using Microsoft.Toolkit.Uwp.Notifications;
-#endif
 #endif
 
 namespace Plugin.LocalNotification
@@ -18,14 +12,13 @@ namespace Plugin.LocalNotification
     /// </summary>
     public static class LocalNotificationExtensions
     {
-#if NET6_0_OR_GREATER
         /// <summary>
         ///
         /// </summary>
         /// <param name="builder"></param>
         /// <param name="configureDelegate"></param>
         /// <returns></returns>
-        public static MauiAppBuilder UseLocalNotification(this MauiAppBuilder builder, Action<ILocalNotificationBuilder> configureDelegate = null)
+        public static MauiAppBuilder UseLocalNotification(this MauiAppBuilder builder, Action<ILocalNotificationBuilder>? configureDelegate = null)
         {
             var localNotificationBuilder = new LocalNotificationBuilder();
             configureDelegate?.Invoke(localNotificationBuilder);
@@ -67,7 +60,7 @@ namespace Plugin.LocalNotification
                 {
                     iOS.FinishedLaunching((application, _) =>
                     {
-                        LocalNotificationCenter.SetUserNotificationCenterDelegate(localNotificationBuilder.IOSBuilder.CustomUserNotificationCenterDelegate);                        
+                        LocalNotificationCenter.SetUserNotificationCenterDelegate(localNotificationBuilder.IOSBuilder.CustomUserNotificationCenterDelegate);
                         return true;
                     });
                     iOS.WillEnterForeground(application =>
@@ -88,6 +81,5 @@ namespace Plugin.LocalNotification
 
             return builder;
         }
-#endif
-            }
+    }
 }
