@@ -1,8 +1,1 @@
-﻿using Foundation;
-
-#if XAMARINIOS
-[assembly: LinkerSafe]
-#elif IOS
-[assembly: System.Reflection.AssemblyMetadata("IsTrimmable", "True")]
-#endif
-
+﻿[assembly: System.Reflection.AssemblyMetadata("IsTrimmable", "True")]
