@@ -1,20 +1,22 @@
 ﻿namespace Plugin.LocalNotification.iOSOption
 {
     /// <summary>
-    /// 
+    ///
     /// </summary>
     public enum iOSLocationAuthorization
     {
         /// <summary>
-        /// 
+        ///
         /// </summary>
         No,
+
         /// <summary>
-        /// 
+        ///
         /// </summary>
         Always,
+
         /// <summary>
-        /// 
+        ///
         /// </summary>
         WhenInUse
     }
